+++ src/XSRenderer/XSImagePNG.cpp
@@ -255,7 +255,7 @@ namespace XS {
 
 			png_bytepp rows = static_cast<png_bytepp>( png_malloc( png, h * sizeof(png_bytep) ) );
 			std::memset( rows, 0, h * sizeof(png_bytep) );
-			for ( size_t i = 0u; i < h; i++ ) {
+			for ( int i = 0; i < h; i++ ) {
 				rows[i] = (png_bytep)(pixels + (h - i - 1) * w * numChannels);
 			}
